Found out that PA6 is broken and getting bad readings.
Changed sensor 5 to use PA11 instead.

With this change, the system meets our C-grade spec of +-18 degrees.

+++ HAB-LDS-AVR/src/state_machine.h
@@ -47,6 +47,6 @@ void main_interrupts_init(void);
 void input_init(void);
 
 // Resolve the angle from the 10 sensor results
-double resolve_angle(double *sensor_results);
+double resolve_angle(int *sensor_results);
 
 #endif
